Move updateNote from noteModal to noteUtils
